Don't show loyalty points payment method on guest checkout
When the user is not logged in, loyalty points payment method should not be
visible (no user, no loyalty points).

This is achieved by overriding the Solidus frontend payment view with Deface
and adding a method in Spree::Order that filters out payments method with
the option preference guest_checkout: false.